+++ cli/commands/user/hash_password.go
@@ -6,6 +6,7 @@ import (
 
 	"github.com/sensu/sensu-go/backend/authentication/bcrypt"
 	"github.com/sensu/sensu-go/cli"
+	"github.com/sensu/sensu-go/cli/commands/hooks"
 	"github.com/spf13/cobra"
 )
 
@@ -29,6 +30,11 @@ func HashPasswordCommand(cli *cli.SensuCli) *cobra.Command {
 
 			fmt.Fprintln(cmd.OutOrStdout(), hash)
 			return nil
+		},
+		Annotations: map[string]string{
+			// We want to be able to run this command regardless of whether the CLI
+			// has been configured.
+			hooks.ConfigurationRequirement: hooks.ConfigurationNotRequired,
 		},
 	}
 }